🐛 fix(database): apply injectSearchSettings consistently for unmodified builtin models
- Fix inconsistency where injectSearchSettings was only called for user-modified models
  but skipped for unmodified builtin models, causing different search toggle behavior
- Remove search ability from LobeHub provider Anthropic models (Bedrock doesn't support
  built-in search)
- Split large aiInfra test file into 7 focused test files under __tests__/
- Add tests for search settings injection consistency
